FEATURE: Resource Creation
Adds ability to create users and tasks

Cleanup:

Removes unused `mongodb.js`
Adds scripts to `package.json`
Creates model files for User and Task and moves them out of
`mongoose.js`
