+++ src/models/user.js
@@ -1,0 +1,42 @@
+const mongoose = require('mongoose');
+const { isEmail } = require('validator');
+
+const User = mongoose.model('User', {
+  name: {
+    type: String,
+    required: true,
+    trim: true
+  },
+  age: {
+    type: Number,
+    default: 0,
+    validate(value) {
+      if (value < 0) {
+        throw new Error('Age cannot be less than 0');
+      }
+    }
+  },
+  email: {
+    type: String,
+    trim: true,
+    lowercase: true,
+    validate(value) {
+      if (!isEmail(value)) {
+        throw new Error('Not a valid email')
+      }
+    }
+  },
+  password: {
+    type: String,
+    minlength: 7,
+    trim: true,
+    required: true,
+    validate(value) {
+      if(value.toLowerCase().includes('password')) {
+        throw new Error('Password cannot include password')
+      }
+    }
+  }
+});
+
+module.exports = User;
